Move `showAndLogExceptionWithTelemetry` out of vscode directory
