+++ extensions/ql-vscode/src/language-support/query-editor.ts
@@ -6,7 +6,7 @@ import { getErrorMessage } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
 import { AppCommandManager, QueryEditorCommands } from "../common/commands";
 import { extLogger } from "../common";
-import { showAndLogExceptionWithTelemetry } from "../common/vscode/logging";
+import { showAndLogExceptionWithTelemetry } from "../common/logging";
 import { telemetryListener } from "../common/vscode/telemetry";
 
 type QueryEditorOptions = {
