+++ extensions/ql-vscode/src/packaging/packaging.ts
@@ -7,12 +7,14 @@ import {
   withProgress,
 } from "../common/vscode/progress";
 import { extLogger } from "../common";
-import { showAndLogExceptionWithTelemetry } from "../common/vscode/logging";
+import {
+  showAndLogExceptionWithTelemetry,
+  showAndLogInformationMessage,
+} from "../common/logging";
 import { asError, getErrorStack } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
 import { PACKS_BY_QUERY_LANGUAGE } from "../common/query-language";
 import { PackagingCommands } from "../common/commands";
-import { showAndLogInformationMessage } from "../common/logging";
 import { telemetryListener } from "../common/vscode/telemetry";
 
 type PackagingOptions = {
