+++ extensions/ql-vscode/src/query-evaluation-logging/eval-log-viewer.ts
@@ -13,7 +13,7 @@ import { asError, getErrorMessage } from "../pure/helpers-pure";
 import { redactableError } from "../pure/errors";
 import { EvalLogViewerCommands } from "../common/commands";
 import { extLogger } from "../common";
-import { showAndLogExceptionWithTelemetry } from "../common/vscode/logging";
+import { showAndLogExceptionWithTelemetry } from "../common/logging";
 import { telemetryListener } from "../common/vscode/telemetry";
 
 export interface EvalLogTreeItem {
